backend/pyproject.toml
[project]
name = "mypay-backend"
version = "1.0.0"
description = "My Pay — Creator Financing Platform API"
requires-python = ">=3.10"

dependencies = [
    # ── Core API ──────────────────────────────────────────────────
    "fastapi==0.110.1",
    "uvicorn==0.25.0",
    "python-multipart>=0.0.9",

    # ── Database ──────────────────────────────────────────────────
    "pymongo==4.5.0",
    "motor==3.3.1",

    # ── Auth / Security ───────────────────────────────────────────
    "pyjwt>=2.10.1",
    "bcrypt==4.1.3",
    "passlib>=1.7.4",
    "cryptography>=42.0.8",
    "pyopenssl>=23.2.0",

    # ── Config / Validation ───────────────────────────────────────
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.0",
    "email-validator>=2.2.0",

    # ── HTTP / Storage ────────────────────────────────────────────
    "requests>=2.31.0",
    "requests-oauthlib>=2.0.0",
    "boto3>=1.34.129",

    # ── Email ─────────────────────────────────────────────────────
    "resend==2.29.0",

    # ── Utilities ─────────────────────────────────────────────────
    "python-jose>=3.3.0",
    "tzdata>=2024.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
